Revert "Merge branch 'dev' into feature/input-validation"
This reverts commit 5140b2a0606076fe8cd5743bf56b0331945dcc9e, reversing
changes made to 0880e0c806ee2df4e6f0882171c972748e21c3db.
